[clean] Moving manifest bits from 'utils' to dedicated 'manifest' module
