+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-01-??
+
+### Changed
+
+- BREAKING: moving the `utils/manifest.py` module into its own `manifest/` module.
+
 ## [0.3.0] - 2023-10-30
 
 ### Added
