+++ src/ledgered/manifest/__init__.py
@@ -1,0 +1,8 @@
+from .app import AppConfig
+from .constants import EXISTING_DEVICES, MANIFEST_FILE_NAME
+from .manifest import LegacyManifest, Manifest
+from .tests import TestsConfig
+
+
+__all__ = ["AppConfig", "EXISTING_DEVICES", "LegacyManifest", "Manifest", "MANIFEST_FILE_NAME", \
+           "TestsConfig"]
