+++ src/ledgered/manifest/constants.py
@@ -1,0 +1,2 @@
+EXISTING_DEVICES = ["nanos", "nanox", "nanos+", "stax"]
+MANIFEST_FILE_NAME = "ledger_app.toml"
